Move Downloader and SQLiteTLE classes inside pyorbital package
